+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+- `Router::getAcceptedResourceType()`
 
 ### Changed
+- Split `Router::getAcceptedType()`
 
 ### Deprecated